winit関連の処理をrendererからuiに移動
