--- src/platform/renderer/mod.rs
@@ -1,94 +1,0 @@
-use std::sync::Arc;
-
-#[allow(unused_imports)]
-use winit::{
-    application::ApplicationHandler,
-    event::*,
-    event_loop::{ActiveEventLoop, EventLoop},
-    keyboard::{KeyCode, PhysicalKey},
-    window::Window,
-};
-
-pub struct State {
-    window: Arc<Window>,
-}
-
-pub struct App {
-    state: Option<State>,
-}
-
-impl State{
-    pub async fn new(window: Arc<Window>) -> anyhow::Result<Self> {
-        Ok(Self {
-            window,
-        })
-    }
-    
-    pub fn resize(&mut self, _width: u32, _height: u32) {
-        // これから実装します。
-    }
-    
-    pub fn render(&mut self) {
-        // ウィンドウにできるだけ早く次のフレームを描画するように要求する
-        self.window.request_redraw();
-    }
-}
-
-#[allow(dead_code)]
-impl App {
-    pub fn new() -> Self {
-        Self {
-            state: None,
-        }
-    }
-}
-
-impl ApplicationHandler<State> for App {
-    fn resumed(&mut self, event_loop: &ActiveEventLoop) {
-        #[allow(unused_mut)]
-        let mut window_attributes = Window::default_attributes();
-
-        let window = Arc::new(event_loop.create_window(window_attributes).unwrap());
-
-        self.state = Some(pollster::block_on(State::new(window)).unwrap());
-
-    }
-
-    #[allow(unused_mut)]
-    fn user_event(&mut self, _event_loop: &ActiveEventLoop, mut event: State) {
-        self.state = Some(event);
-    }
-
-    fn window_event(
-        &mut self,
-        event_loop: &ActiveEventLoop,
-        _window_id: winit::window::WindowId,
-        event: WindowEvent,
-    ) {
-        let state = match &mut self.state {
-            Some(canvas) => canvas,
-            None => return,
-        };
-
-        match event {
-            WindowEvent::CloseRequested => event_loop.exit(),
-            WindowEvent::Resized(size) => state.resize(size.width, size.height),
-            WindowEvent::RedrawRequested => {
-                state.render();
-            }
-            WindowEvent::KeyboardInput {
-                event:
-                    KeyEvent {
-                        physical_key: PhysicalKey::Code(code),
-                        state: key_state,
-                        ..
-                    },
-                ..
-            } => match (code, key_state.is_pressed()) {
-                (KeyCode::Escape, true) => event_loop.exit(),
-                _ => {}
-            },
-            _ => {}
-        }
-    }
-}
